Avoid an overflow in roundToIntegral in obscure formats, fix CVC4 bug 2932
If the extra bit(s) added when going from the packed to the unpacked
significand can't represent 2*significand then this overflows.  Given the
bits are chosen so that they can represent at least significand extra bits,
this only affects formats with very small significands and proportionally
(exponentially?) large signficands.  3/5 is one of these.