Executable test vectors for Aina's bug report symfpu #4
This was reported in https://github.com/martin-cs/symfpu/issues/4

It is caused by bug in the symfpu-0.0.7-CVC4-patched branch that
was used by CVC4/5 and Bitwuzla for a while.  The code and the
bug didn't ever make it to main which solves the problem a
different way, in part to avoid bugs like this!

For completeness and to prevent regressions, I added this to the
executable test vectors.