Fix precondition of `conditionalNegate()`
`conditionalNegate()` was asserting that if we negate the input, then
the input cannot correspond to `100...0`. However, this is problematic
when the operation is used conditionally. This is for example necessary
in `convertFloatToSBV()`, where we want to apply the negation if the
result is not the safe overflow value. This commit removes the assertion
and adds a comment to `conditionalNegate()` that documents why we cannot
have an assertion there.

Additionally, the commit makes `convertFloatToSBV()` safer by explicitly
handling the safe overflow case. Previously, we negated such a value,
even though it cannot be negated. This worked out in the cvc5 case
(because it blindly computes `~x + 1` without asserting anything), but
could be dangerous.